I lied
Cant stop wont stop
Fixed most if not all the rest of the bugs for the testcases given.
NOTE: '#' are handled just by returning null, i dont think there is anything telling us to make it a token so i guess its ignored, started a function "checkOther" where ive been throwing problems.
For some reason on eustice '\n' is not the same as a carriage return so i added ASCII  13 to check other